requirement.txt
langchain==0.1.20 
langchain-community==0.0.38 
langchain-openai==0.1.6 
langchain-text-splitters
chromadb 
sentence-transformers 
docx2txt 
python-dotenv 
streamlit 
neo4j